requirements.txt
# ============================================================================
# requirements.txt - Dependencias del proyecto
# ============================================================================

playwright==1.41.0
pandas==2.1.4
supabase==2.3.0
python-dotenv==1.0.0